Added Intro Pages
The outest navigator is now "InitialNavigator". It nests: (1) the old "RightsTabNavigator" (which in turn nests "RightsNavigator"), and (2) a new SwitchNavigator that contains the intro pages.

TODO: Implement auto-scrolling on the language selection's "next" button, add a gear icon somewhere in the app to navigate back to the intro pages in case user wants to change language / role, and dynamic-ify the array of languages (future).